Move network graph functions into separate file
