+++ src/components/Play.js
@@ -3,6 +3,7 @@ import {Helmet} from 'react-helmet';
 import Sticky from 'react-stickynode';
 import {Link} from 'react-router-dom';
 import api from '../api';
+import {makeGraph} from '../network';
 import PlayDetailsNav from './PlayDetailsNav';
 import IdLink from './IdLink';
 import Years from './Years';
@@ -31,66 +32,6 @@ const navItems = [
 
 const tabNames = new Set(navItems.map(item => item.name));
 
-function getCooccurrences (segments) {
-  const map = {};
-  segments.forEach(s => {
-    if (!s.speakers) {
-      return;
-    }
-
-    // make sure each speaker occurs only once in scene
-    const speakers = s.speakers.filter((v, i, a) => a.indexOf(v) === i);
-    speakers.forEach((c, i) => {
-      if (i < speakers.length - 1) {
-        const others = speakers.slice(i + 1);
-        others.forEach(o => {
-          const pair = [c, o].sort();
-          const key = pair.join('|');
-          if (map[key]) {
-            map[key][2]++;
-          } else {
-            map[key] = pair.concat(1);
-          }
-        });
-      }
-    });
-  });
-
-  const cooccurrences = [];
-  Object.keys(map)
-    .sort()
-    .forEach(key => {
-      cooccurrences.push(map[key]);
-    });
-
-  return cooccurrences;
-}
-
-function makeGraph (cast, segments) {
-  const nodes = [];
-  cast.forEach(p => {
-    nodes.push({id: p.id, label: p.name || `#${p.id}`});
-  });
-  const cooccurrences = getCooccurrences(segments);
-  const edges = [];
-  cooccurrences.forEach(cooc => {
-    edges.push({
-      id: cooc[0] + '|' + cooc[1],
-      source: cooc[0],
-      target: cooc[1],
-      size: cooc[2],
-      // NB: we set the edge color here since the defaultEdgeColor in Sigma
-      // settings does not to have any effect
-      color: edgeColor
-    });
-  });
-  return {nodes, edges};
-}
-
-// TODO: refactor to reduce complexity
-// see https://eslint.org/docs/rules/complexity
-/* eslint "complexity": ["error", 30] */
-
 const PlayInfo = ({corpusId, playId}) => {
   const [play, setPlay] = useState(null);
   const [graph, setGraph] = useState(null);
@@ -105,7 +46,7 @@ const PlayInfo = ({corpusId, playId}) => {
         const response = await api.get(url);
         if (response.ok) {
           const {cast, segments} = response.data;
-          const graph = makeGraph(cast, segments);
+          const graph = makeGraph(cast, segments, edgeColor);
           setPlay(response.data);
           setGraph(graph);
         } else if (response.status === 404) {
